feat(sprite): extract collision detection
