+++ kontra.mjs
@@ -2325,42 +2325,6 @@ vectorFactory.prototype = Vector.prototype;
 vectorFactory.class = Vector;
 
 /**
- * A collection of collision detection functions.
- *
- * @sectionName Collision
- */
-
-/**
- *
- * @function spriteCollidesWith
- *
- * @param {Sprite} sprite - Sprite object.
- * @param {Object} object - Object to check collision against.
- *
- * @returns {Boolean|null} `true` if the objects collide, `false` otherwise. Will return `null` if the either of the two objects are rotated.
- */
-
-function spriteCollidesWith(sprite, object) {
-  if (sprite.rotation || object.rotation) return null;
-
-  // take into account sprite anchors
-  let x = sprite.x - sprite.width * sprite.anchor.x;
-  let y = sprite.y - sprite.height * sprite.anchor.y;
-
-  let objX = object.x;
-  let objY = object.y;
-  if (object.anchor) {
-    objX -= object.width * object.anchor.x;
-    objY -= object.height * object.anchor.y;
-  }
-
-  return x < objX + object.width &&
-         x + sprite.width > objX &&
-         y < objY + object.height &&
-         y + sprite.height > objY;
-}
-
-/**
  * A versatile way to update and draw your game objects. It can handle simple rectangles, images, and sprite sheet animations. It can be used for your main player object as well as tiny particles in a particle engine.
  * @class Sprite
  *
@@ -2743,35 +2707,8 @@ class Sprite {
   }
 
   /**
-   * Check if the sprite collide with the object. Uses a simple [Axis-Aligned Bounding Box (AABB) collision check](https://developer.mozilla.org/en-US/docs/Games/Techniques/2D_collision_detection#Axis-Aligned_Bounding_Box). Takes into account the sprites [anchor](api/sprite/#anchor).
-   *
-   * **NOTE:** Does not take into account sprite rotation. If you need collision detection between rotated sprites you will need to implement your own `collidesWith()` function. I suggest looking at the Separate Axis Theorem.
-   *
-   * ```js
-   * import { Sprite } from 'kontra';
-   *
-   * let sprite = Sprite({
-   *   x: 100,
-   *   y: 200,
-   *   width: 20,
-   *   height: 40
-   * });
-   *
-   * let sprite2 = Sprite({
-   *   x: 150,
-   *   y: 200,
-   *   width: 20,
-   *   height: 20
-   * });
-   *
-   * sprite.collidesWith(sprite2);  //=> false
-   *
-   * sprite2.x = 115;
-   *
-   * sprite.collidesWith(sprite2);  //=> true
-   * ```
-   *
-   * If you need a different type of collision check, you can override this function by passing an argument by the same name.
+   * Check if the sprite collides with the object. Actually this functions always returns false but it is declared here for documentation and plugins registration.
+   * You can check the [Collision](api/collision) section for basic collision detection functions.
    *
    * ```js
    * // circle collision
@@ -2804,10 +2741,10 @@ class Sprite {
    *
    * @param {Object} object - Object to check collision against.
    *
-   * @returns {Boolean|null} `true` if the objects collide, `false` otherwise. Will return `null` if the either of the two objects are rotated.
+   * @returns {Boolean|null} `false` always.
    */
   collidesWith(object) {
-    return spriteCollidesWith(this, object);
+    return false;
   }
 
   /**
@@ -3875,6 +3812,71 @@ function TileEngine(properties = {}) {
   return tileEngine;
 }
 
+/**
+ * A collection of collision detection functions.
+ *
+ * @sectionName Collision
+ */
+
+/**
+ * Check if the sprite collide with the object. Uses a simple [Axis-Aligned Bounding Box (AABB) collision check](https://developer.mozilla.org/en-US/docs/Games/Techniques/2D_collision_detection#Axis-Aligned_Bounding_Box). Takes into account the sprites [anchor](api/sprite/#anchor).
+ *
+ * **NOTE:** Does not take into account sprite rotation. If you need collision detection between rotated sprites, I suggest looking at the Separate Axis Theorem.
+ *
+ * ```js
+ * import { Sprite, spriteCollidesWith } from 'kontra';
+ *
+ * let sprite = Sprite({
+ *   x: 100,
+ *   y: 200,
+ *   width: 20,
+ *   height: 40,
+ *   collidesWith: function (object) {
+ *     return spriteCollidesWith(this, object);
+ *   },
+ * });
+ *
+ * let sprite2 = Sprite({
+ *   x: 150,
+ *   y: 200,
+ *   width: 20,
+ *   height: 20
+ * });
+ *
+ * sprite.collidesWith(sprite2);  //=> false
+ *
+ * sprite2.x = 115;
+ *
+ * sprite.collidesWith(sprite2);  //=> true
+ * ```
+ * @function spriteCollidesWith
+ *
+ * @param {Sprite} sprite - Sprite base object.
+ * @param {Object} object - Object to check collision against.
+ *
+ * @returns {Boolean|null} `true` if the objects collide, `false` otherwise. Will return `null` if the either of the two objects are rotated.
+ */
+
+function spriteCollidesWith(sprite, object) {
+  if (sprite.rotation || object.rotation) return null;
+
+  // take into account sprite anchors
+  let x = sprite.x - sprite.width * sprite.anchor.x;
+  let y = sprite.y - sprite.height * sprite.anchor.y;
+
+  let objX = object.x;
+  let objY = object.y;
+  if (object.anchor) {
+    objX -= object.width * object.anchor.x;
+    objY -= object.height * object.anchor.y;
+  }
+
+  return x < objX + object.width &&
+         x + sprite.width > objX &&
+         y < objY + object.height &&
+         y + sprite.height > objY;
+}
+
 let kontra = {
   Animation: animationFactory,
 
@@ -3888,6 +3890,8 @@ let kontra = {
   loadAudio,
   loadData,
   load,
+
+  spriteCollidesWith,
 
   init,
   getCanvas,
