+++ test/unit/sprite.spec.js
@@ -477,7 +477,7 @@ describe('sprite', () => {
   // --------------------------------------------------
   describe('collidesWith', () => {
 
-    it('should correctly detect collision between two objects', () => {
+    it('always returns false', () => {
       let sprite1 = Sprite({
         x: 10,
         y: 20,
@@ -492,99 +492,11 @@ describe('sprite', () => {
         height: 20
       });
 
-      expect(sprite1.collidesWith(sprite2)).to.be.true;
-
-      sprite2.x = 10;
-      sprite2.y = 20;
-
-      expect(sprite1.collidesWith(sprite2)).to.be.true;
-
-      sprite2.x = 1;
-      sprite2.y = 1;
-
-      expect(sprite1.collidesWith(sprite2)).to.be.true;
-
-      sprite2.x = 20;
-      sprite2.y = 40;
-
       expect(sprite1.collidesWith(sprite2)).to.be.false;
-
-      sprite2.x = 0;
-      sprite2.y = 0;
-
-      expect(sprite1.collidesWith(sprite2)).to.be.false;
-    });
-
-    it('should take into account sprite.anchor', () => {
-      let sprite1 = Sprite({
-        x: 10,
-        y: 20,
-        width: 10,
-        height: 20,
-        anchor: {
-          x: 0.5,
-          y: 0.5
-        }
-      });
-
-      let sprite2 = Sprite({
-        x: 10,
-        y: 20,
-        width: 10,
-        height: 20
-      });
-
-      expect(sprite1.collidesWith(sprite2)).to.be.true;
-
-      sprite1.anchor = {x: 1, y: 0};
-
-      expect(sprite1.collidesWith(sprite2)).to.be.false;
-
-      sprite2.anchor = {x: 1, y: 0};
-
-      expect(sprite1.collidesWith(sprite2)).to.be.true;
-    });
-
-    it('should return null if either sprite is rotated', () => {
-      let sprite1 = Sprite({
-        x: 10,
-        y: 20,
-        width: 10,
-        height: 20,
-        rotation: 1
-      });
-
-      let sprite2 = Sprite({
-        x: 19,
-        y: 39,
-        width: 10,
-        height: 20
-      });
-
-      expect(sprite1.collidesWith(sprite2)).to.equal(null);
-
-      sprite1.rotation = 0;
-      sprite2.rotation = 1;
-
-      expect(sprite1.collidesWith(sprite2)).to.equal(null);
-    });
-
-    it('should work for non-sprite objects', () => {
-      let sprite1 = Sprite({
-        x: 10,
-        y: 20,
-        width: 10,
-        height: 20
-      });
-
-      let obj = {
-        x: 10,
-        y: 20,
-        width: 10,
-        height: 20
-      };
-
-      expect(sprite1.collidesWith(obj)).to.be.true;
+      expect(sprite1.collidesWith(sprite1)).to.be.false;
+      expect(sprite1.collidesWith({})).to.be.false;
+      expect(sprite1.collidesWith(undefined)).to.be.false;
+      expect(sprite1.collidesWith(null)).to.be.false;
     });
 
   });
